requirements.txt
asab
motor
numpy
simplejson
imutils
aiohttp-jinja2
nanocamera
hurry.filesize
aiohttp-cors==0.7.0
